docker/functest/functional_test/requirements.txt
# requirements.txt v1.0
# ---------------------
# Add project specific python requirements to this file.
# Do not commit them in the project!
# Make sure they exist on our corporate PyPi server.

pyhamcrest==1.8.0
pytz>=2014
pytest==4.4.1
mock==1.0.1
dnspython==1.14.0
boto==2.48.0
future==0.17.0
requests==2.20.0
pytest-xdist==1.29.0
